fs: consolidate read/write logic, check bit functions
Create functions to set, unset, and check bitmap index before reading
or writing. This is a step towards create a commit/rollback system to
atomize each kind of file operation.

As we consolidate read/writes, we will eventually only set bits during
the allocation phase of each block/inode, allocate a list of "changes"
as we walk through an operation, then commit all of those changes once
we complete the logic. As we commit changes, we'll also collect the
previous information at each step in case we need to revert those
changes, which we'll do if anything wrong happens.