fs: add read/write bits for file pointer state
Note: we will DEFINITELY want to reorganize junk from this and the next
couple of commits. Right now "fdtable" is serving as the house for what
should be file pointer objects (fdtable_entry) and the table itself
(fdtable). Really, there should be a separate module for this, but
right now fs.h has these constants (and likely fs.c will handle a lot
of the read/write logic) and fdtable handles anything directly
connected to the table and its entries.