fs: fix swap deletion in dir_rm
If the swapped file is any of the existing pointers (the "next,"
"last_prev", or "prev" pointers), then these should be treated as
special cases, otherwise the pointers will be miscalculated (or the
wrong calculations will be applied to certain pointers).

Additionally, perform the swap first so the re-link of prev to next can
use the latest information.