fs_ata: create module
fs.c will be for all the high-level logic within the public API. fs_ata
is a self-contained module that performs read/write operations in the
context of the filesystem. It would, for instance, set particular bits
in a bitmap and write "blocks" consisting of multiple physical blocks
on the disk. It is an interface to ata for use with the fs_* functions.