fs_transaction: make commit_list_init/append void
There's really nothing we can check that would make the transaction
unable to function. kfree has a limited ability to check for problems
because even an incorrect pointer can still point to invalid memory.
Then, even if an error is identified, it doesn't mean that the
transaction object is unusable, as we are starting with an empty list.