fs_transaction: add rollback mechanism
Create function try_commit, which will contain logic of original
commit, and rollback, which will revert the changes of a commit if any
particular action fails.

One of the important decisions here is to make all of the data members
value-based. This is to avoid complications around allocating heap
memory--if we coan allocate the whole transaction in the first place,
we shouldn't have to worry about anything else. The transaction's
rollback actions are created in place, replacing the commit that has
been completed.

Based on immediate testing, it seems like there will have to be some
changes to when a commit is marked as complete, or even a distinction
called is_revertable, indicating that we have read all the data to
replace the original. Right now, there are still some instances where a
read of existing data can fail after modification occurred.