string: prevent integer overflow in printf_putint
Instead of letting f (the place value iterator) get larger than the
i (the subject to be printed), stop the loop when f is at exactly the
largest place value the number has, not one place value larger. This
prevents f from overflowing, causing the function to print the number
incorrectly.